# appointment_api_swagger.yaml
swagger: "2.0"
info:
  title: "Appointment API"
  description: "API for managing provider availability and scheduling patient appointments."
  version: "1.0.0"
host: "localhost:5002" # Default from appointment_api.py
basePath: "/api"
schemes:
  - "http" # For development
  # - "https" # For production
consumes:
  - "application/json"
produces:
  - "application/json"
definitions:
  Error:
    type: "object"
    description: "Standard error response structure."
    required:
      - status
      - message
    properties:
      status:
        type: "string"
        enum: ["error"]
        example: "error"
      message:
        type: "string"
        description: "A human-readable description of the error."
        example: "Missing required fields."
...
  AppointmentRequestPayload:
    type: "object"
    description: "Payload for a patient to request a new appointment."
    required:
      - patient_id
      - provider_id
      - appointment_start_time
      - appointment_end_time
    properties:
      patient_id:
        type: "integer"
        format: "int32"
        description: "ID of the patient requesting the appointment."
        example: 101
      provider_id:
        type: "integer"
        format: "int32"
        description: "ID of the provider for the appointment."
        example: 1
      appointment_start_time:
        type: "string"
        format: "date-time"
        description: "Requested start time for the appointment (YYYY-MM-DD HH:MM:SS)."
        example: "2024-08-01 10:00:00"
      appointment_end_time:
        type: "string"
        format: "date-time"
        description: "Requested end time for the appointment (YYYY-MM-DD HH:MM:SS)."
        example: "2024-08-01 10:30:00"
      reason_for_visit:
        type: "string"
        nullable: true
        description: "Optional reason for the visit."
        example: "Annual Checkup"
      notes_by_patient:
        type: "string"
        nullable: true
        description: "Optional notes from the patient."
        example: "Prefer morning slots if possible."
...
paths:
...
  /appointments/request:
    post:
      summary: "Request an Appointment"
      description: "Allows a patient to request a new appointment with a provider."
      operationId: "requestAppointment"
      tags: ["Appointments"]
      parameters:
        - name: "body"
          in: "body"
          required: true
          description: "Details of the appointment request."
          schema:
            $ref: "#/definitions/AppointmentRequestPayload"
      responses:
        "201":
          description: "Appointment requested successfully."
          schema:
            type: "object"
            properties:
              status: { type: "string", example: "success" }
              message: { type: "string", example: "Appointment requested successfully. Awaiting provider confirmation." }
              appointment_id: { type: "integer", format: "int32" }
        "400":
          description: "Bad Request (e.g., missing fields, invalid user IDs, invalid time range)."
          schema: { $ref: "#/definitions/Error" }
        "500":
          description: "Internal Server Error."
          schema: { $ref: "#/definitions/Error" }
...